Add files via upload
This is a .py file that can be created in the terminal and run in in the command line using $python3 myquast.py <file.fa> 
The purpose is to provide an output similar to that of QUAST, summarizing a fasta file and providing the following four outputs: N50 value, total number of contigs, total length of all contigs, and length of the largest contig. 
In this example the input file is "contigs.fa" - this code can be run on any fasta file format.